Capture memory usage in the status page
Running things in Cloud Foundry means that observability isn’t quite
where we’d like it to be. `psutil` in Python doesn’t work.

This change is an experiment to see whether we get meaningful data out
of Go applications. If it succeeds, then we might be able to use a
similar mechanism (in terms of C system calls) to improve what we have
for Python applications.